move theme Types to their own class replace logger with dialog
