+++ src/main/java/org/jabref/gui/preferences/general/GeneralTabViewModel.java
@@ -21,6 +21,7 @@ import org.jabref.gui.DialogService;
 import org.jabref.gui.desktop.JabRefDesktop;
 import org.jabref.gui.preferences.PreferenceTabViewModel;
 import org.jabref.gui.theme.Theme;
+import org.jabref.gui.theme.ThemeTypes;
 import org.jabref.gui.util.DirectoryDialogConfiguration;
 import org.jabref.gui.util.FileDialogConfiguration;
 import org.jabref.logic.l10n.Language;
@@ -39,25 +40,8 @@ import de.saxsys.mvvmfx.utils.validation.FunctionBasedValidator;
 import de.saxsys.mvvmfx.utils.validation.ValidationMessage;
 import de.saxsys.mvvmfx.utils.validation.ValidationStatus;
 import de.saxsys.mvvmfx.utils.validation.Validator;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 public class GeneralTabViewModel implements PreferenceTabViewModel {
-    protected enum ThemeTypes {
-        LIGHT(Localization.lang("Light")),
-        DARK(Localization.lang("Dark")),
-        CUSTOM(Localization.lang("Custom..."));
-
-        private final String displayName;
-
-        ThemeTypes(String displayName) {
-            this.displayName = displayName;
-        }
-
-        public String getDisplayName() {
-            return displayName;
-        }
-    }
 
     protected static SpinnerValueFactory<Integer> fontSizeValueFactory =
             new SpinnerValueFactory.IntegerSpinnerValueFactory(9, Integer.MAX_VALUE);
@@ -66,8 +50,6 @@ public class GeneralTabViewModel implements PreferenceTabViewModel {
             new ReadOnlyListWrapper<>(FXCollections.observableArrayList(Language.values()));
     private final ObjectProperty<Language> selectedLanguageProperty = new SimpleObjectProperty<>();
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(GeneralTabViewModel.class);
-    
     private final ReadOnlyListProperty<ThemeTypes> themesListProperty =
             new ReadOnlyListWrapper<>(FXCollections.observableArrayList(ThemeTypes.values()));
     private final ObjectProperty<ThemeTypes> selectedThemeProperty = new SimpleObjectProperty<>();
@@ -329,11 +311,10 @@ public class GeneralTabViewModel implements PreferenceTabViewModel {
 
     public void openBrowser() {
         String url = "https://themes.jabref.org";
-
         try {
             JabRefDesktop.openBrowser(url, preferences.getFilePreferences());
         } catch (java.io.IOException e) {
-            LOGGER.error("An IOException occurred:", e);
+            dialogService.showErrorDialogAndWait(Localization.lang("Could not open browser."));
         }
     }
 }
