+++ src/main/java/org/jabref/gui/theme/ThemeTypes.java
@@ -1,0 +1,20 @@
+package org.jabref.gui.theme;
+
+import org.jabref.logic.l10n.Localization;
+
+public enum ThemeTypes {
+
+        LIGHT(Localization.lang("Light")),
+        DARK(Localization.lang("Dark")),
+        CUSTOM(Localization.lang("Custom..."));
+
+        private final String displayName;
+
+        ThemeTypes(String displayName) {
+            this.displayName = displayName;
+        }
+
+        public String getDisplayName() {
+            return displayName;
+        }
+}
